executor/cte_test.go: remove suite, increase parallelism
